+++ lib/mesa/src/virtio/vulkan/vn_device_memory.h
@@ -31,7 +31,7 @@ struct vn_device_memory {
    struct vn_renderer_bo *base_bo;
    /* enforce kernel and ring ordering between memory export and free */
    bool bo_roundtrip_seqno_valid;
-   uint32_t bo_roundtrip_seqno;
+   uint64_t bo_roundtrip_seqno;
    VkDeviceSize base_offset;
 
    VkDeviceSize map_end;
